# Conflicts:
#	backend/cdkproxymain.py
#	backend/dataall/base/cdkproxy/cdk_cli_wrapper.py
#	backend/dataall/base/cdkproxy/requirements.txt
#	backend/dataall/base/cdkproxy/stacks/__init__.py
#	backend/dataall/base/loader.py
#	backend/dataall/core/environment/api/resolvers.py
#	backend/dataall/core/environment/cdk/pivot_role_stack.py
#	backend/dataall/core/environment/services/environment_service.py
#	backend/dataall/core/permissions/permissions.py
#	backend/dataall/modules/catalog/api/__init__.py
#	backend/dataall/modules/catalog/api/mutations.py
#	backend/dataall/modules/catalog/api/queries.py
#	backend/dataall/modules/catalog/api/resolvers.py
#	backend/dataall/modules/catalog/api/types.py
#	backend/dataall/modules/catalog/db/glossary_repositories.py
#	backend/dataall/modules/dashboards/__init__.py
#	backend/dataall/modules/dashboards/api/resolvers.py
#	backend/dataall/modules/dashboards/indexers/dashboard_indexer.py
#	backend/dataall/modules/dashboards/services/dashboard_service.py
#	backend/dataall/modules/datapipelines/__init__.py
#	backend/dataall/modules/datapipelines/api/queries.py
#	backend/dataall/modules/datapipelines/api/resolvers.py
#	backend/dataall/modules/datapipelines/api/types.py
#	backend/dataall/modules/datapipelines/aws/codecommit_datapipeline_client.py
#	backend/dataall/modules/datapipelines/cdk/datapipelines_cdk_cli_wrapper_extension.py
#	backend/dataall/modules/datapipelines/cdk/datapipelines_cdk_pipeline.py
#	backend/dataall/modules/datapipelines/cdk/datapipelines_pipeline.py
#	backend/dataall/modules/datapipelines/db/datapipelines_repositories.py
#	backend/dataall/modules/datapipelines/services/datapipelines_service.py
#	backend/dataall/modules/dataset_sharing/aws/ram_client.py
#	backend/dataall/modules/dataset_sharing/db/share_object_repositories.py
#	backend/dataall/modules/dataset_sharing/services/share_permissions.py
#	backend/dataall/modules/datasets/__init__.py
#	backend/dataall/modules/datasets/api/dataset/resolvers.py
#	backend/dataall/modules/datasets/api/storage_location/resolvers.py
#	backend/dataall/modules/datasets/api/table/resolvers.py
#	backend/dataall/modules/datasets/aws/athena_table_client.py
#	backend/dataall/modules/datasets/aws/glue_profiler_client.py
#	backend/dataall/modules/datasets/aws/lf_dataset_client.py
#	backend/dataall/modules/datasets/aws/sns_dataset_client.py
#	backend/dataall/modules/datasets/cdk/assets/gluedatabasecustomresource/index.py
#	backend/dataall/modules/datasets/cdk/assets/lakeformationdefaultsettings/index.py
#	backend/dataall/modules/datasets/cdk/dataset_custom_resources_extension.py
#	backend/dataall/modules/datasets/cdk/dataset_stack.py
#	backend/dataall/modules/datasets/indexers/dataset_indexer.py
#	backend/dataall/modules/datasets/services/dataset_location_service.py
#	backend/dataall/modules/datasets/services/dataset_permissions.py
#	backend/dataall/modules/datasets/services/dataset_service.py
#	backend/dataall/modules/datasets/services/dataset_table_service.py
#	backend/dataall/modules/datasets_base/__init__.py
#	backend/dataall/modules/feed/api/mutations.py
#	backend/dataall/modules/feed/api/resolvers.py
#	backend/dataall/modules/feed/api/types.py
#	backend/dataall/modules/mlstudio/__init__.py
#	backend/dataall/modules/mlstudio/cdk/__init__.py
#	backend/dataall/modules/mlstudio/cdk/mlstudio_stack.py
#	backend/dataall/modules/mlstudio/services/mlstudio_permissions.py
#	backend/dataall/modules/vote/api/resolvers.py
#	backend/dataall/modules/vote/db/vote_repositories.py
#	backend/dataall/modules/worksheets/__init__.py
#	backend/dataall/modules/worksheets/services/worksheet_permissions.py
#	backend/docker/dev/Dockerfile
#	backend/local_graphql_server.py
#	backend/migrations/versions/4a0618805341_rename_sgm_studio_permissions.py
#	backend/migrations/versions/5fc49baecea4_add_enviromental_parameters.py
#	backend/requirements.txt
#	config.json
#	deploy/cdk_exec_policy/cdkExecPolicy.yaml
#	deploy/stacks/auth_at_edge.py
#	deploy/stacks/backend_stack.py
#	deploy/stacks/container.py
#	deploy/stacks/pipeline.py
#	deploy/stacks/run_if.py
#	docker-compose.yaml
#	frontend/docker/dev/Dockerfile
#	frontend/package-lock.json
#	frontend/package.json
#	frontend/src/design/components/layout/DefaultSidebar.js
#	frontend/src/modules/Environments/components/EnvironmentFeatures.js
#	frontend/src/modules/Environments/views/EnvironmentView.js
#	frontend/src/modules/Omics/views/OmicsRunCreateForm.js
#	frontend/src/modules/Worksheets/views/WorksheetView.js
#	frontend/src/utils/helpers/moduleUtils.js
#	frontend/yarn.lock
#	tests/conftest.py
#	tests/modules/catalog/test_glossary.py
#	tests/modules/datapipelines/test_datapipelines.py
#	tests/modules/datasets/conftest.py
#	tests/modules/datasets/test_dataset_feed.py
#	tests/modules/datasets/test_dataset_glossary.py
#	tests/modules/datasets/test_dataset_location.py
#	tests/modules/mlstudio/cdk/test_sagemaker_studio_stack.py

+++ backend/dataall/core/environment/services/environment_service.py
@@ -275,22 +275,14 @@ class EnvironmentService:
 
     @staticmethod
     def validate_permissions(session, uri, g_permissions, group):
+        """
+        g_permissions: coming from frontend = ENVIRONMENT_INVITATION_REQUEST
+
+        """
         if permissions.INVITE_ENVIRONMENT_GROUP in g_permissions:
-            g_permissions.append(permissions.LIST_ENVIRONMENT_GROUPS)
             g_permissions.append(permissions.REMOVE_ENVIRONMENT_GROUP)
 
-        if permissions.ADD_ENVIRONMENT_CONSUMPTION_ROLES in g_permissions:
-            g_permissions.append(permissions.LIST_ENVIRONMENT_CONSUMPTION_ROLES)
-
-        if permissions.CREATE_NETWORK in g_permissions:
-            g_permissions.append(permissions.LIST_ENVIRONMENT_NETWORKS)
-
-        g_permissions.append(permissions.GET_ENVIRONMENT)
-        g_permissions.append(permissions.LIST_ENVIRONMENT_GROUPS)
-        g_permissions.append(permissions.LIST_ENVIRONMENT_GROUP_PERMISSIONS)
-        g_permissions.append(permissions.LIST_ENVIRONMENT_NETWORKS)
-        g_permissions.append(permissions.CREDENTIALS_ENVIRONMENT)
-
+        g_permissions.extend(permissions.ENVIRONMENT_INVITED_DEFAULT)
         g_permissions = list(set(g_permissions))
 
         if g_permissions not in permissions.ENVIRONMENT_INVITED:
@@ -330,6 +322,15 @@ class EnvironmentService:
             raise exceptions.EnvironmentResourcesFound(
                 action='Remove Team',
                 message=f'Team: {group} has created {group_env_objects_count} resources on this environment.',
+            )
+
+        group_env_consumption_roles = EnvironmentService.query_user_environment_consumption_roles(
+            session, [group], uri, {}
+        ).all()
+        if group_env_consumption_roles:
+            raise exceptions.EnvironmentResourcesFound(
+                action='Remove Team',
+                message=f'Team: {group} has consumption role(s) on this environment.',
             )
 
         group_membership = EnvironmentService.find_environment_group(
@@ -814,32 +815,51 @@ class EnvironmentService:
             .filter(EnvironmentGroup.environmentUri == uri)
             .all()
         )
-        for group in env_groups:
-
-            session.delete(group)
-
-            ResourcePolicy.delete_resource_policy(
-                session=session,
-                resource_uri=uri,
-                group=group.groupUri,
-            )
-
         env_roles = (
             session.query(ConsumptionRole)
             .filter(ConsumptionRole.environmentUri == uri)
             .all()
         )
+
+        env_resources = 0
+        for group in env_groups:
+            env_resources += EnvironmentResourceManager.count_group_resources(
+                session,
+                environment,
+                group.groupUri
+            )
         for role in env_roles:
-            session.delete(role)
+            env_resources += EnvironmentResourceManager.count_consumption_role_resources(
+                session,
+                role.consumptionRoleUri
+            )
 
-        KeyValueTag.delete_key_value_tags(
-            session, environment.environmentUri, 'environment'
-        )
+        if env_resources > 0:
+            raise exceptions.EnvironmentResourcesFound(
+                action='Delete Environment',
+                message=f'Found {env_resources} resources on environment {environment.label} - Delete all environment related objects before proceeding',
+            )
+        else:
+            EnvironmentResourceManager.delete_env(session, environment)
+            EnvironmentParameterRepository(session).delete_params(environment.environmentUri)
 
-        EnvironmentResourceManager.delete_env(session, environment)
-        EnvironmentParameterRepository(session).delete_params(environment.environmentUri)
+            for group in env_groups:
+                session.delete(group)
 
-        return session.delete(environment)
+                ResourcePolicy.delete_resource_policy(
+                    session=session,
+                    resource_uri=uri,
+                    group=group.groupUri,
+                )
+
+            for role in env_roles:
+                session.delete(role)
+
+            KeyValueTag.delete_key_value_tags(
+                session, environment.environmentUri, 'environment'
+            )
+
+            return session.delete(environment)
 
     @staticmethod
     def get_environment_parameters(session, env_uri):
